backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bcrypt==5.0.0
blis==1.3.3
catalogue==2.0.10
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
cloudpathlib==0.23.0
confection==0.1.5
cryptography==46.0.3
cymem==2.0.13
ecdsa==0.19.1
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl
exceptiongroup==1.3.1
fastapi==0.128.0
h11==0.16.0
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
lxml==6.0.2
MarkupSafe==3.0.3
murmurhash==1.0.15
numpy==2.0.2
packaging==26.0
passlib==1.7.4
preshed==3.0.12
pyasn1==0.6.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pypdf==6.6.2
python-docx==1.2.0
python-jose==3.5.0
python-multipart==0.0.20
requests==2.32.5
rsa==4.9.1
scikit-learn==1.6.1
scipy==1.13.1
six==1.17.0
smart_open==7.5.0
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.46
sqlmodel==0.0.31
srsly==2.5.2
starlette==0.49.3
thinc==8.3.9
threadpoolctl==3.6.0
tqdm==4.67.1
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.39.0
wasabi==1.1.3
weasel==0.4.3
wrapt==2.0.1
